set_massless_SIS_tracers default to compute domain
The `set_massless_SIS_tracers` function would default to the data domain
for its mass fields, even when

In production this had little effect, since `mass(:,:,:)` would often
default to zero and would generally only affect unused values in the
halos, but would raise errors in more aggressive initialization tests.

This patch modifies the default behavior to use the compute grid.  The
data grid can now be used if 1) the flag is present and 2) it is set to
false.

In all likelihood, this flag is no longer needed, and may be purged in
the future.